ai-service/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
redis>=5.0.0
httpx>=0.28.1
aiohttp>=3.9.0
# LLM Integration
groq>=0.4.0
openai>=1.0.0
